[server] move routes config to routes.go
